Refactor organization hooks to useDollyOrganisasjoner across multiple components
