+++ apps/dolly-frontend/src/main/js/src/components/bestilling/statusListe/BestillingResultat/BestillingResultat.tsx
@@ -8,7 +8,7 @@ import './BestillingResultat.less'
 import GjenopprettConnector from '@/components/bestilling/gjenopprett/GjenopprettBestillingConnector'
 import useBoolean from '@/utils/hooks/useBoolean'
 import { REGEX_BACKEND_GRUPPER, useMatchMutate } from '@/utils/hooks/useMutate'
-import { Bestillingsstatus } from '@/utils/hooks/useOrganisasjoner'
+import { Bestillingsstatus } from '@/utils/hooks/useDollyOrganisasjoner'
 import { BestillingStatus } from '@/components/bestilling/statusListe/BestillingProgresjon/BestillingStatus'
 import { TestComponentSelectors } from '#/mocks/Selectors'
 import ConfettiExplosion from 'react-confetti-explosion'
