+++ apps/dolly-frontend/src/main/js/src/components/fagsystem/inntektsmelding/form/partials/orgnrToogle.tsx
@@ -5,7 +5,10 @@ import { OrganisasjonToggleGruppe } from '@/components/organisasjonSelect/Organi
 import { EgneOrganisasjoner, getEgneOrganisasjoner } from '@/utils/EgneOrganisasjoner'
 import { UseFormReturn } from 'react-hook-form/dist/types'
 import { useCurrentBruker } from '@/utils/hooks/useBruker'
-import { useDollyFasteDataOrganisasjoner, useOrganisasjoner } from '@/utils/hooks/useOrganisasjoner'
+import {
+	useDollyFasteDataOrganisasjoner,
+	useDollyOrganisasjoner,
+} from '@/utils/hooks/useDollyOrganisasjoner'
 import { ArbeidsgiverTyper } from '@/components/fagsystem/aareg/AaregTypes'
 import Loading from '@/components/ui/loading/Loading'
 import { getOrgType } from '@/utils/OrgUtils'
@@ -29,7 +32,7 @@ export const OrgnrToggle = ({
 		useDollyFasteDataOrganisasjoner()
 
 	const { organisasjoner: brukerOrganisasjoner, loading: brukerOrganisasjonerLoading } =
-		useOrganisasjoner(currentBruker?.brukerId)
+		useDollyOrganisasjoner(currentBruker?.brukerId)
 	const egneOrganisasjoner = getEgneOrganisasjoner(brukerOrganisasjoner)
 
 	const orgnr = formMethods.watch(virksomhetPath)
