+++ apps/dolly-frontend/src/main/js/src/pages/organisasjoner/Organisasjoner.tsx
@@ -11,7 +11,7 @@ import { dollySlack } from '@/components/dollySlack/dollySlack'
 import TomOrgListe from './TomOrgliste'
 import { useNavigate } from 'react-router'
 import { useCurrentBruker } from '@/utils/hooks/useBruker'
-import { useOrganisasjonBestilling } from '@/utils/hooks/useOrganisasjoner'
+import { useOrganisasjonBestilling } from '@/utils/hooks/useDollyOrganisasjoner'
 import { sokSelector } from '@/ducks/bestillingStatus'
 import { useDispatch } from 'react-redux'
 import { resetPaginering } from '@/ducks/finnPerson'
